verizon.lua: update to responde on status codes
Abort when getting a 0 status code. Abort after 5 tries when getting status code above 500 or status code above 400 except 404 and 403 on a verizon.net or bellatlantic.net site. Do nothing ater 5 tries when getting status code above 500 or status code above 400 except 404 and 403 on a site that is not verizon.net or bellatlantic.net.